rename: Restore the filename of CalendarModal.tsx
