+++ components/ModalContainer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useModalContext } from '../contexts/ModalContext';
 
-import CalendarModal from './calendarModaltemp';
+import CalendarModal from './CalendarModal';
 import UserModal from './UserModal';
 
 export const enum MODAL_NAMES {
